feat(example): the database is declared, and the container follows
The example is where the paradigm has to be legible, and it still had the
triplication the design exists to remove: a DatabaseService parsing
DATABASE_URL, `services: { postgres: true }` for the container, and an
env file restating the URL a third time.

All three become one line. `KyselyDatabase<Database, 'Example'>` is what
implies a Postgres, names the database inside it, and provides
EXAMPLE_URL; endpoints reach it as `db` through the router rather than as
`services.database`. Nothing lists postgres locally any more -- the entry
that survives in `docker.compose.services` is the deploy half, ignored by
reconcile until deploy derives from the manifest too.

EXAMPLE_URL is deliberately absent from `.env.example`: env files are
loaded after secrets, so restating a derived URL overrides the port that
was actually allocated. Migrations run through `gkm exec` for the same
reason -- it is where the injection happens.

`profile.ts` called `db.query`, which Kysely has no such method for. It
is a `sql` template now, which also parameterises the user id it used to
interpolate into the statement.
