+++ apps/example/.env.example
@@ -1,2 +1,4 @@
 TEST_VAR=1
-DATABASE_URL=postgresql://geekmidas:geekmidas@localhost:5432/examples
+# EXAMPLE_URL is deliberately absent: the `Example` database construct declares
+# it, and `gkm dev` / `gkm exec` inject it from the container they reconciled.
+# Restating it here would override the port that was actually allocated.
