+++ apps/example/src/config/env.ts
@@ -14,9 +14,12 @@ export const config = envParser
 		port: get('PORT').string().transform(Number).default(3000),
 		nodeEnv: get('NODE_ENV').string().default('development'),
 		database: {
-			url: get('DATABASE_URL')
+			// Provided by the `Example` database construct — `gkm dev` and
+			// `gkm exec` inject it from the container they reconciled, so the port
+			// is whatever was free rather than whatever was hard-coded.
+			url: get('EXAMPLE_URL')
 				.string()
-				.default('postgresql://geekmidas:geekmidas@localhost:5432/examples'),
+				.default('postgres://geekmidas:geekmidas@localhost:5432/example'),
 		},
 	}))
 	.parse();
