+++ apps/example/src/endpoints/router.ts
@@ -1,8 +1,8 @@
 import { e } from '@geekmidas/constructs/endpoints';
 import logger from '../config/logger.js';
+import { database } from '../constructs/database';
 import { AuditStorageService } from '../services/AuditStorageService';
 import { AuthService } from '../services/AuthService';
-import { DatabaseService } from '../services/DatabaseService';
 import { EventsService } from '../services/EventsService';
 
 /**
@@ -11,16 +11,16 @@ import { EventsService } from '../services/EventsService';
  * The pattern:
  * ```typescript
  * e.logger(logger)
- *   .services([DatabaseService, AuthService])
- *   .database(DatabaseService)
+ *   .services([AuthService])
+ *   .database(database.service)
  *   .auditor(AuditStorageService)
  *   .publisher(EventsService)
  * ```
  *
  * Creates a factory where all endpoints inherit:
  * - The same logger instance
- * - Access to registered services (database, auth, etc.)
- * - Database connection available as `db` in handler context
+ * - Access to registered services (auth, etc.)
+ * - The declared database, available as `db` in handler context
  * - Auditor available in handler context for manual audits
  * - Support for declarative audits via `.audit([...])`
  *
@@ -32,14 +32,14 @@ import { EventsService } from '../services/EventsService';
  *   .handle(async ({ services, logger, db, auditor }) => {
  *     // db - database connection (or transaction when audits are atomic)
  *     // auditor - for manual audit logging
- *     // services.database, services.auth - registered services
+ *     // services.auth - registered services
  *   });
  * ```
  */
 export const router = e
 	.logger(logger)
-	.services([DatabaseService, AuthService])
-	.database(DatabaseService)
+	.services([AuthService])
+	.database(database.service)
 	.auditor(AuditStorageService)
 	.publisher(EventsService)
 	.authorizer('iam');
